Add files via upload
Final update to the stand-alone GUI that was developed to test husky and it's features. This version does not include the battery feature Jamie created. This version does support keyboard movement and input values that control speed. There may be comments on the code, ignore them as I was trying to figure out the battery feature as well.